+++ SchoolManagementSystem/Commands/Teachers/DeleteTeachersCommand.cs
@@ -1,6 +1,7 @@
 ﻿using SchoolManagementSystem.Models;
 using SchoolManagementSystem.Services.Interface;
 using SchoolManagementSystem.ViewModels.UserControls;
+using SchoolManagementSystem.Views.Windows;
 using System;
 using System.Collections.Generic;
 using System.Collections.ObjectModel;
@@ -23,6 +24,11 @@ namespace SchoolManagementSystem.Commands.Teachers
 
         public override void Execute(object parameter)
         {
+            SureDialogWindow dialogWindow = new SureDialogWindow();
+            dialogWindow.ShowDialog();
+            if (dialogWindow.DialogResult != true)
+                return;
+
             int id = _viewModel.SelectedValue.Id;
 
             _teacherService.Delete(id);
